refactored and formatted code
