+++ pom.xml
@@ -112,9 +112,9 @@
 				</configuration>
 			</plugin>
 			<plugin>
-				<groupId>com.coveo</groupId>
+				<groupId>com.spotify.fmt</groupId>
 				<artifactId>fmt-maven-plugin</artifactId>
-				<version>2.9</version>
+				<version>2.20</version>
 				<executions>
 					<execution>
 						<goals>
